% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_icon_labels.R
\name{add_axis_icons}
\alias{add_axis_icons}
\title{xxx}
\usage{
add_axis_icons(breaks, labels, country_icons, logo_icons, width, ...)
}
\arguments{
\item{xx}{xxx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
fill
}
\keyword{xxx}
