% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{box_plot}
\alias{box_plot}
\title{Boxplot ggOxford}
\usage{
box_plot(
  data,
  group_level = "Cohort 1",
  strata_name = "Overall",
  strata_level = "Overall",
  variable = "age"
)
}
\arguments{
\item{data}{dataset}

\item{group_level}{group_level in dataset}

\item{strata_name}{strata_name in dataset}

\item{strata_level}{strata_level in dataset}

\item{variable}{variable that the box plot should plot over}
}
\value{
a box plot
}
\description{
Boxplot ggOxford
}
