% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forestPlot}
\alias{forestPlot}
\title{Forest Plot - Oxford-Pharmacoepi style}
\usage{
forestPlot(data, outcomelabels, estimate, lowCi, highCi)
}
\arguments{
\item{data:}{Include your data frame.}

\item{"outcome_labels":}{name of your column including the  labels. It has to go within " ".}

\item{"estimate":}{include name of the column with RR, IR, HR, etc. It has to go within " ".}

\item{"lowCI":}{include name of the column with the low 95\% confidence interval value. It has to go within " ".}

\item{"highCI":}{include name of the column with the high 95\% confidence interval value. It has to go within " ".}
}
\description{
Forest Plot - Oxford-Pharmacoepi style
}
