% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_icon_labels.R
\name{geom_axis_flags}
\alias{geom_axis_flags}
\title{Add X-Axis flags to ggplots}
\usage{
geom_axis_flags(
  breaks,
  labels,
  country_icons,
  width = 50,
  lineheight = 3,
  fontface = "bold",
  ...
)
}
\arguments{
\item{breaks}{x axis breaks}

\item{labels}{vector of labels to depict}

\item{country_icons}{country for which flag to depict, in capital letter ISO-3 format (e.g. DEU for Germany)}

\item{width}{width of the flag, defaults to 50}

\item{lineheight}{height of the axis label line}

\item{fontface}{Font Face for Labels, defaults to "bold"}
}
\value{
adds x axis flags to ggplot
}
\description{
Add X-Axis flags to ggplots
}
\note{
You must not call a ggplot theme after the call of geom_axis_flags, otherwise the theme parameter necessary for flag rendering will be overwritten.
}
\examples{
options(tidyverse.quiet = TRUE)
library(tidyverse)
library(ggoxford)
library(ggtext)
wpop2013 <-
  tidyr::population \%>\%
  filter(year == 2013) \%>\%
  filter(population > 1e6) \%>\%
  mutate(iso3 = countrycode::countrycode(country,
                                         origin = "country.name.en",
                                         destination = "iso3c")
  ) \%>\%
  slice(1:10)
head(wpop2013,10)
ggplot(data = wpop2013 ,
       aes(x = iso3, y = population)) +
  geom_bar(stat = "identity") +
  theme_minimal() +
  geom_axis_flags(breaks = wpop2013$iso3,
                   labels = wpop2013$country,
                   country_icons = wpop2013$iso3,
                   width = 30,
                   lineheight = 2,
                   fontface = "bold"
  )

}
\references{
Flags are under CC-BY licence and are obtained from: https://github.com/eosrei/emojione-color-font
}
