% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBarChart.R
\name{getBarChart}
\alias{getBarChart}
\title{getBarChart summary}
\usage{
getBarChart(tab, cohort, var = c("age_group"), color = "grey")
}
\arguments{
\item{tab}{dataframe}

\item{cohort}{cohort}

\item{var}{variable that is going to be for the bar chart}

\item{color}{color}
}
\value{
plot
}
\description{
getBarChart summary
}
