% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggoxford-package.R
\docType{package}
\name{ggoxford-package}
\alias{ggoxford}
\alias{ggoxford-package}
\title{ggoxford: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
