% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockForestPlotData.R
\name{mockForestPlotData}
\alias{mockForestPlotData}
\title{Function to create mock forest plot data.}
\usage{
mockForestPlotData(seed = 1, numVars = 8)
}
\arguments{
\item{seed}{Seed value for the random mock data generated}

\item{numVars}{Number of variables}
}
\description{
Function to create mock forest plot data.
}
\examples{
\donttest{
library(ggoxford)

mockForestPlotData()
}

}
