% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockSummarisedResult.R
\name{mockSummarisedResult}
\alias{mockSummarisedResult}
\title{Function to create a mock summarised result tibble.}
\usage{
mockSummarisedResult(seed = 1, populationsize = 5)
}
\arguments{
\item{seed}{Seed value for the random mock data generated}

\item{populationsize}{choose size of the population}
}
\description{
Function to create a mock summarised result tibble.
}
\examples{
\donttest{
library(ggoxford)

mockSummarisedResult()
}

}
