% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockSummariseLargeScaleCharacteristics.R
\name{mockSummariseLargeScaleCharacteristics}
\alias{mockSummariseLargeScaleCharacteristics}
\title{Function to create mock summarised large-scale characteristics.}
\usage{
mockSummariseLargeScaleCharacteristics(
  populationSize = 10,
  window = list(c(-Inf, -366), c(-365, -31), c(-30, -1), c(1, 30), c(31, 365), c(366,
    Inf)),
  eventInWindow = "condition_occurrence",
  episodeInWindow = "drug_exposure"
)
}
\arguments{
\item{populationSize}{Number of individuals}

\item{window}{Set of temporal windows to consider}

\item{eventInWindow}{Table to characterise the events in the window}

\item{episodeInWindow}{Table to characterise the episodes in the window}
}
\description{
Function to create mock summarised large-scale characteristics.
}
\examples{
\donttest{
library(ggoxford)

mockSummariseLargeScaleCharacteristics()
}

}
