% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxford-colours.R
\docType{data}
\name{oxblue}
\alias{oxblue}
\title{Arial Narrow font name R variable aliases}
\format{
length 1 character vector
}
\usage{
oxblue
}
\description{
\code{font_an} == "\verb{Arial Narrow}"
}
\keyword{datasets}
