% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBarChart.R
\name{plotBarChart}
\alias{plotBarChart}
\title{getBarChart summary}
\usage{
plotBarChart(
  data,
  variable,
  groupName = "cohort_name",
  groupLevel = "Cohort 1",
  strataName = "Overall",
  strataLevel = "Overall"
)
}
\arguments{
\item{data}{Input data.}

\item{variable}{Variable to prepare the barchart.}

\item{groupName}{Group name to use.}

\item{groupLevel}{Group level to use.}

\item{strataName}{Strata name to use.}

\item{strataLevel}{Strata levels to use.}
}
\value{
plot
}
\description{
getBarChart summary
}
