% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeOxford.R
\name{themeOxford}
\alias{themeOxford}
\title{Function to create a ggplot theme}
\usage{
themeOxford(plot, baseSize = 8)
}
\arguments{
\item{plot}{ggplot}

\item{baseSize}{base font size, given in pts.}
}
\value{
ggplot
}
\description{
Function to create a ggplot theme
}
