% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBarChart.R
\name{getBarChart}
\alias{getBarChart}
\title{getBarChart summary}
\usage{
getBarChart(tab, var)
}
\arguments{
\item{tab}{dataframe}

\item{var}{variable that is going to be for the bar chart}
}
\value{
plot
}
\description{
getBarChart summary
}
