% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockSummarisedResult.R
\name{mockSummarisedResult}
\alias{mockSummarisedResult}
\title{Function to create a mock summarised result tibble.}
\usage{
mockSummarisedResult(seed = 1, start_date, end_date)
}
\arguments{
\item{seed}{Seed value for the random mock data generated}

\item{start_date}{Start date for the mock data}

\item{end_date}{End date for the mock data}
}
\description{
Function to create a mock summarised result tibble.
}
\examples{
\donttest{
library(ggoxford)

mockSummarisedResult(seed = 1,as.Date("2021-01-01"),as.Date("2021-12-31"))
}

}
